LayoutTests/fast/element-targeting/target-container-with-visual-overflow.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1">
<script src="../../resources/ui-helper.js"></script>
<script src="../../resources/js-test.js"></script>
<style>
body, html {
    width: 100%;
    height: 100%;
    margin: 0;
    font-size: 18px;
    font-family: system-ui;
    text-align: right;
}

.visible-overflow {
    width: 100%;
    height: 150px;
    text-align: left;
    border-bottom: 1px solid lightgray;
}

.main-content {
    width: 100%;
    height: 100vh;
}

.target {
    width: 160px;
    height: 150px;
    display: inline-block;
    text-align: center;
    color: white;
    line-height: 150px;
    border-radius: 6px;
    margin-right: 4px;
}

.red {
    background: red;
}

.green {
    background: green;
}
</style>
<script>
jsTestIsAsync = true;
addEventListener("load", async () => {
    selector = await UIHelper.adjustVisibilityForFrontmostTarget(50, 50);
    shouldBeEqualToString("selector", "#expected");
    finishJSTest();
});
</script>
</head>
<body>
    <div class="visible-overflow">
        <div class="main-content">
            <span id="expected" class="green target">Right</span>
            <span class="red target">Wrong</span>
        </div>
        <span class="red target">Wrong</span>
    </div>
    <span class="red target">Wrong</span>
    <p id="console"></p>
</body>
</html>
